+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.4.2- 2021-05-27
+
+- Use another parser
+- Revert h function change
+- Use cache for setReactivity
+
 ## 1.4.1- 2021-05-21
 
 - Use better RegExp